Wrap event state handling
The initial fetching of events is far from ideal, since they're being
preloaded for all rooms, but it works for now.
I'm also unsure how great it is to register a separate event listener
for each room 🤔 It could make sense to consolidate that into a single
listener that pushes events to their respective rooms.

closes #3