Add cancel button to loading view
Sometimes the SDK gets into a weird state where we have already saved
invalid credentials, but can't log in due to it thinking the server
URL is broken. This way we can get out of that without having to clear
the app's keychain on launch.
Also it's nice to have.