Move context menu all the way down the msg view chain
Now it displays only on the message bubble, not the reactions view,
which was kinda broken with the long press on that. Another plus is
that the flipped view isn't as big anymore.
Second plus is that it's no longer possible to open the context menu
on redacted messages possibly redacting them again or sending
reactions. Probably better that way.