Remove incorrect reaction assertion
Reactions should ideally only be single emoji, but they're not always.
This is especially apparent when briding to somewhere with custom
emoji like Slack. Matrix clients will just have to deal with the fact
that some reactions will be something like `:this:`.

Also, the `isSingleEmoji` property apparently dies on some actual
single emoji, which is awkward.